chore(avm): L5 - rename lookup_into_indexed_by_clk.hpp to match its contents
The file defines LookupIntoIndexedByRow and is included via the
LookupIntoIndexedByRow enum value, but the filename still referenced
the old "clk" naming. Rename it to lookup_into_indexed_by_row.hpp.
