+++ barretenberg/cpp/src/barretenberg/vm2/tracegen/lib/lookup_into_indexed_by_row.hpp
